feat(resolve): add ApiDOM YAML 1.2 parser
Refs #2717
