--- src/helpers/apidom/openapi-3-1/parsers/json/index.js
@@ -1,48 +1,0 @@
-import { from, ParseResultElement } from '@swagger-api/apidom-core';
-import { ParserError } from '@swagger-api/apidom-reference/configuration/empty';
-
-const JsonParser = {
-  name: 'json-swagger-client',
-  fileExtensions: [],
-  mediaTypes: ['application/json'],
-
-  async canParse(file) {
-    const hasSupportedFileExtension =
-      this.fileExtensions.length === 0 ? true : this.fileExtensions.includes(file.extension);
-    const hasSupportedMediaType = this.mediaTypes.includes(file.mediaType);
-
-    if (!hasSupportedFileExtension) return false;
-    if (hasSupportedMediaType) return true;
-    if (!hasSupportedMediaType) {
-      try {
-        JSON.parse(file.toString());
-        return true;
-      } catch (e) {
-        return false;
-      }
-    }
-    return false;
-  },
-
-  async parse(file) {
-    if (this.sourceMap) {
-      // eslint-disable-next-line no-console
-      console.warn("json-swagger-client parser plugin doesn't support sourceMaps option");
-    }
-
-    const source = file.toString();
-
-    try {
-      const element = from(JSON.parse(source));
-      const parseResultElement = new ParseResultElement();
-
-      element.classes.push('result');
-      parseResultElement.push(element);
-      return parseResultElement;
-    } catch (error) {
-      throw new ParserError(`Error parsing "${file.uri}"`, error);
-    }
-  },
-};
-
-export default JsonParser;
